Added to iRacingConnection object three fields detailing performance metrics for the main data loop.
ProcessingTime: The time it takes to read and deserialise the data from
the shared memory buffer with iracing.

WaitingTime: The time it takes to wait for data to be ready

YieldTime: The time the processing of the user's main loop took.

All times are in nanoseconds.

See the sample SamplePerformanceMeasurement for more details.