Incidents: work around incident scanning issue
Appears to be a bug, where iracing does not respond correctly
to the 'MoveToNextIncident' message.  It simply moves 5 seconds.
Same thing for 'MoveToPreviousIncident'.  It does not change cam car.

It appears iRacing enter into a state, whereby the incident scanning fails -
initially its working - until iRacing for unknown reasons fails to response
to the incident scanning messages correctly.

Fix seems to be to select the pace as camcar first, then advance to next incident
If camcar is still on then pace car - we can then assume not an incident

The side effect of doing this, seems to avoid iRacing entering into the buggy state!

It appears that